[funexpected] Update Rect2, Refactor
- add more methods for Rect2
- put implementation into separate files
- change include structure
